Import shared TAR helpers from common.utils
